web-apps(front): add TranslationModule to common for jupyter
