+++ components/crud-web-apps/common/frontend/kubeflow-common-lib/package.json
@@ -28,6 +28,8 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.26",
     "@fortawesome/free-brands-svg-icons": "^5.12.0",
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
+    "@ngx-translate/core": "^12.0.0",
+    "@ngx-translate/http-loader": "^4.0.0",
     "date-fns": "1.29.0",
     "lodash-es": "4.17.14",
     "material-icons": "^0.3.1",
