+++ components/crud-web-apps/common/frontend/kubeflow-common-lib/projects/kubeflow/src/lib/form/name-namespace-inputs/name-namespace-inputs.component.html
@@ -8,7 +8,7 @@
   ></lib-name-input>
 
   <mat-form-field appearance="outline" class="column">
-    <mat-label>Namespace</mat-label>
+    <mat-label>{{ "formName.lblNamespace" | translate }}</mat-label>
     <input
       matInput
       placeholder="Namespace"
